requirements.txt
Package                       Version
----------------------------- ---------
asgiref                       3.4.1
certifi                       2022.9.24
cffi                          1.15.1
charset-normalizer            2.0.12
coreapi                       2.3.3
coreschema                    0.0.4
cryptography                  38.0.3
defusedxml                    0.7.1
Django                        3.2.16
django-cors-headers           3.10.1
django-templated-mail         1.1.1
djangorestframework           3.14.0
djangorestframework-simplejwt 4.4.0
djoser                        2.1.0
idna                          3.4
importlib-metadata            1.7.0
itypes                        1.2.0
Jinja2                        3.0.3
MarkupSafe                    2.0.1
oauthlib                      3.2.2
Pillow                        8.4.0
pip                           21.3.1
pkg_resources                 0.0.0
pycparser                     2.21
PyJWT                         1.7.1
python3-openid                3.2.0
pytz                          2022.4
requests                      2.27.1
requests-oauthlib             1.3.1
setuptools                    39.0.1
six                           1.16.0
social-auth-app-django        4.0.0
social-auth-core              4.0.2
sqlparse                      0.4.3
typing_extensions             4.1.1
uritemplate                   4.1.1
urllib3                       1.26.12
zipp                          3.6.0
